kitt/requirements.txt
livekit-agents~=0.7.2
livekit-plugins-openai~=0.6
livekit-plugins-deepgram~=0.5
livekit-plugins-elevenlabs~=0.6
livekit-plugins-silero~=0.5
aiohttp
httpx
